Rename some files and variables
